DCS-76 Extracting out report summary object to reuse on a couple of different 'view report' pages
